* Fixed PL_get_integer() to check for overflow

+++ src/pl-fli.c
@@ -691,14 +691,14 @@ PL_get_integer(term_t t, int *i)
   }
   if ( isReal(w) )
   { real f = valReal(w);
-    long l;
+    int l;
 
 #ifdef DOUBLE_TO_LONG_CAST_RAISES_SIGFPE
     if ( !((f >= PLMININT) && (f <= PLMAXINT)) )
       fail;
 #endif
 
-    l = (long)f;
+    l = (int)f;
     if ( (real)l == f )
     { *i = l;
       succeed;
